Time/Date changes for SQL
sqlite supports 3 different date formats, but doesn't have a specific
date type.

We are going to use (for now) TEXT as the type and ISO-8601 as the
format. These changes introduce iso-8601 as the format for booking date
time.

Thankfully, the date-fns lib is very nice.

Dropping support for release date time.