Update inmate table to use time types
Converted dob and booking_date to date and timestampz respectively.
Performed testing. This conversion was so simple and elegant. Exactly
what I wanted! All thanks to IS0-8601 format & SQLITE's preference for
that.

Its easy to 'group by date(booking_date)' now! I should change the
analytics to use postgres for the calculations and not js.